mod softmax, logSoftmax
