+++ tfjs-core/src/ops/ops.ts
@@ -77,6 +77,7 @@ export {leakyRelu} from './leaky_relu';
 export {less} from './less';
 export {lessEqual} from './less_equal';
 export {localResponseNormalization} from './local_response_normalization';
+export {logSoftmax} from './log_softmax';
 export {logSumExp} from './log_sum_exp';
 export {logicalAnd} from './logical_and';
 export {logicalNot} from './logical_not';
@@ -132,17 +133,20 @@ export {slice1d} from './slice1d';
 export {slice2d} from './slice2d';
 export {slice3d} from './slice3d';
 export {slice4d} from './slice4d';
+export {softmax} from './softmax';
 export {spaceToBatchND} from './space_to_batch_nd';
 export {split} from './split';
 export {square} from './square';
 export {squaredDifference} from './squared_difference';
 export {squeeze} from './squeeze';
 export {stack} from './stack';
+export {stridedSlice} from './strided_slice';
 export {sub} from './sub';
 export {sum} from './sum';
 export {tan} from './tan';
 export {tanh} from './tanh';
 export {tile} from './tile';
+export {topk} from './topk';
 export {truncatedNormal} from './truncated_normal';
 export {unsortedSegmentSum} from './unsorted_segment_sum';
 export {unstack} from './unstack';
@@ -156,11 +160,8 @@ export * from './binary_ops';
 export * from './array_ops';
 export * from './tensor_ops';
 export * from './transpose';
-export * from './softmax';
 export * from './norm';
 export * from './moving_average';
-export * from './strided_slice';
-export * from './topk';
 export * from './scatter_nd';
 export * from './spectral_ops';
 export * from './sparse_to_dense';
