+++ tfjs-core/src/ops/softmax.ts
@@ -16,9 +16,10 @@
  */
 
 import {ENGINE} from '../engine';
-import {customGrad} from '../gradients';
+import {Softmax, SoftmaxAttrs, SoftmaxInputs} from '../kernel_names';
+import {NamedAttrMap} from '../kernel_registry';
 import {Tensor} from '../tensor';
-import {GradSaveFunc} from '../tensor_types';
+import {NamedTensorMap} from '../tensor_types';
 import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 
@@ -56,8 +57,8 @@ function softmax_<T extends Tensor>(logits: T|TensorLike, dim = -1): T {
         `Logits was rank ${$logits.rank} and dim was ${dim}`);
   }
 
-  const inputsToSave: Tensor[] = [];
-  const outputsToSave = [true];
+  const inputs: SoftmaxInputs = {logits: $logits};
+  const attrs: SoftmaxAttrs = {dim};
 
   return ENGINE.runKernelFunc(
       (backend, save) => {
@@ -65,69 +66,8 @@ function softmax_<T extends Tensor>(logits: T|TensorLike, dim = -1): T {
         save([y]);
         return y;
       },
-      {logits: $logits},
-      (dy: T, saved: Tensor[]) => {
-        const [y] = saved;
-        const dyTimesY = dy.mul(y);
-        const keepDims = true;
-
-        return {
-          logits: () => dyTimesY.sub(dyTimesY.sum([dim], keepDims).mul(y))
-        };
-      },
-      'Softmax', {dim}, inputsToSave, outputsToSave);
-}
-
-/**
- * Computes the log softmax.
- *
- * ```js
- * const a = tf.tensor1d([1, 2, 3]);
- *
- * a.logSoftmax().print();  // or tf.logSoftmax(a)
- * ```
- *
- * ```js
- * const a = tf.tensor2d([2, 4, 6, 1, 2, 3], [2, 3]);
- *
- * a.logSoftmax().print();  // or tf.logSoftmax(a)
- * ```
- *
- * @param logits The logits array.
- * @param axis The dimension softmax would be performed on. Defaults to `-1`
- *     which indicates the last dimension.
- */
-/** @doc {heading: 'Operations', subheading: 'Normalization'} */
-function logSoftmax_<T extends Tensor>(logits: T|TensorLike, axis = -1): T {
-  const $logits = convertToTensor(logits, 'logits', 'logSoftmax');
-
-  if (axis === -1) {
-    axis = $logits.rank - 1;
-  }
-  if (axis !== $logits.rank - 1) {
-    throw Error(
-        'Log Softmax along a non-last dimension is not yet supported. ' +
-        `Logits was rank ${$logits.rank} and axis was ${axis}`);
-  }
-
-  const customOp = customGrad((logits: Tensor, save: GradSaveFunc) => {
-    const keepDims = true;
-    const xMax = logits.max(axis, true);
-    const shifted = logits.sub(xMax);
-    const value =
-        shifted.toFloat().sub(shifted.exp().sum(axis, keepDims).log());
-    save([value]);
-    const gradFunc = (dy: T, saved: Tensor[]) => {
-      const [value] = saved;
-      const softmax = value.exp();
-      return dy.sub(dy.sum(axis, keepDims).mul(softmax));
-    };
-
-    return {value, gradFunc};
-  });
-
-  return customOp($logits) as T;
+      inputs as {} as NamedTensorMap, null /* grad */, Softmax,
+      attrs as {} as NamedAttrMap);
 }
 
 export const softmax = op({softmax_});
-export const logSoftmax = op({logSoftmax_});
